Moves types.ts to src folder
