+++ src/components/OUSByIslandCard.tsx
@@ -16,7 +16,7 @@ import {
   getPrecalcMetrics,
   toPercentMetric,
 } from "../../data/bin/getPrecalcMetrics";
-import { GeoProp } from "../util/types";
+import { GeoProp } from "../types";
 import { getGeographyById } from "../util/getGeographyById";
 
 interface ByIslandProp extends GeoProp {
