+++ src/functions/boundaryAreaOverlap.ts
@@ -28,7 +28,7 @@ import {
   clipSketchToGeography,
   getParamStringArray,
 } from "../util/clipSketchToGeography";
-import { ExtraParams } from "../util/types";
+import { ExtraParams } from "../types";
 
 const metricGroup = project.getMetricGroup("boundaryAreaOverlap");
 const boundaryTotalMetrics = getPrecalcMetrics(
